requirements.txt
cryptography
python-dotenv
scapy
